Fix the :app compile errors, and add the tool that found them (#4)
`:app` had never been compiled - Google's Maven is unreachable from the
environment it was written in, so AGP, Compose, CameraX and Room cannot be
resolved and :app cannot even be configured there.

Rather than fix the reported error alone and wait for the next one, this adds
tools/typecheck: a standalone build that compiles the whole :app source tree
against Compose Multiplatform's Maven Central artifacts plus stubs for the
Android surface. It found 14 further errors beyond the one reported.

Fixed:

  - Charts.kt used the pointer-input `size` (IntSize) where the DrawScope's
    `Size` was meant, passing Int to a Float parameter. This was the reported
    error.
  - TopAppBar is @ExperimentalMaterial3Api in Material3 1.3, so all 11 call
    sites failed. Opted in at file level across the seven screen files.
  - LiveScreen assigned `preview.surfaceProvider`; CameraX 1.4 has no getter,
    so there is no such property. Calls setSurfaceProvider instead.
  - SpeechAnnouncer assigned `tts.language`; setLanguage returns an int status,
    so Kotlin exposes no setter. Calls setLanguage instead.
  - Three more icons still used the non-auto-mirrored forms, deprecated in
    Material Icons: ShowChart, OpenInNew, Undo.

:app now type-checks with no errors and no warnings. That is not the same as
building: AGP, Room's generated code and everything at runtime remain
unverified, and Compose Multiplatform is close to but not identical with the
Android artifacts. tools/typecheck/README.md is specific about the gap, and
about why a wrong stub signature is worse than a missing one.

:core and :vision still pass 177 tests.

+++ app/src/main/java/com/nobrainsoft/rangeanalyser/ui/common/Charts.kt
@@ -84,7 +84,8 @@ fun TrendChart(
                     // Tap to inspect: a phone has no hover, and a value on every point would be
                     // unreadable.
                     detectTapGestures { tap ->
-                        val plot = Plot(size.width, size.height, trend)
+                        // `size` here is the pointer-input IntSize, not the DrawScope's Size.
+                        val plot = Plot(size.width.toFloat(), size.height.toFloat(), trend)
                         selected = plot.nearestIndex(tap.x)
                     }
                 },
